EL-3617 - Enable release of beta versions of UX Aspects (#9)
* Changes to support release of beta versions

The SEPG Auto Release job will not release a build containing a prerelease affix, e.g. `2.0.0-beta.3`. This proposed workaround will be to 1) set the main pom.xml version of each project to `2.0.0-SNAPSHOT` (remove the current prerelease specifier) and 2) specify the optional `VERSION_PRERELEASE` environment variable within the pom.xml file containing the required prerelease affix (e.g. `beta.3`).

This will allow the Jenkins job to process the build as a 2.0.0 release. In the build script, instead of taking the environment VERSION directly, we will combine this with the `VERSION_PRERELEASE` value when populating the version in package.json, causing the NPM package to contain the full version including prerelease.

* Also support for CI releases when beta.x is removed from the pom version