Sanitize unsafe characters in the version; fix for beta builds
Added a normalization step to functions which return version identifiers potentially including a prerelease. This addresses the issue where SEPG injects an underscore into version prereleases.

Added `getVersionForCI` for use in setversion. This no longer applies the explicit prerelease (e.g. "beta.6") to the CI version. This was causing trouble when building related branches together, because the beta handling stuff is a workaround that SEPG doesn't know anything about. For now the CI builds will be `2.0.0-SNAPSHOT` instead of `2.0.0-beta.6-SNAPSHOT`, so I think it will be safe as long as we don't have to work on two prereleases of the same library at once.